api/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements/requirements.txt
#
alembic==1.8.1
    # via data-inclusion-api (setup.py)
anyio==3.6.2
    # via
    #   httpcore
    #   starlette
    #   watchfiles
certifi==2022.9.24
    # via
    #   httpcore
    #   httpx
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
click==8.1.3
    # via
    #   data-inclusion-api (setup.py)
    #   uvicorn
cryptography==38.0.3
    # via python-jose
data-inclusion-schema==0.7.0
    # via data-inclusion-api (setup.py)
dnspython==2.2.1
    # via email-validator
ecdsa==0.18.0
    # via python-jose
email-validator==1.3.0
    # via pydantic
faker==15.3.1
    # via data-inclusion-api (setup.py)
fastapi==0.86.0
    # via
    #   data-inclusion-api (setup.py)
    #   fastapi-pagination
    #   sentry-sdk
fastapi-pagination==0.10.0
    # via data-inclusion-api (setup.py)
greenlet==2.0.1
    # via sqlalchemy
gunicorn==20.1.0
    # via data-inclusion-api (setup.py)
h11==0.12.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.15.0
    # via httpx
httptools==0.5.0
    # via uvicorn
httpx==0.23.0
    # via data-inclusion-api (setup.py)
idna==3.4
    # via
    #   anyio
    #   email-validator
    #   requests
    #   rfc3986
mako==1.2.3
    # via alembic
markupsafe==2.1.1
    # via mako
psycopg2==2.9.5
    # via data-inclusion-api (setup.py)
pyasn1==0.4.8
    # via
    #   python-jose
    #   rsa
pycparser==2.21
    # via cffi
pydantic[email]==1.10.2
    # via
    #   data-inclusion-api (setup.py)
    #   data-inclusion-schema
    #   fastapi
    #   fastapi-pagination
python-dateutil==2.8.2
    # via faker
python-dotenv==0.21.0
    # via
    #   data-inclusion-api (setup.py)
    #   uvicorn
python-jose[cryptography]==3.3.0
    # via data-inclusion-api (setup.py)
pytz==2022.6
    # via data-inclusion-api (setup.py)
pyyaml==6.0
    # via uvicorn
requests==2.28.1
    # via data-inclusion-api (setup.py)
rfc3986[idna2008]==1.5.0
    # via httpx
rsa==4.9
    # via python-jose
sentry-sdk[fastapi]==1.11.1
    # via data-inclusion-api (setup.py)
six==1.16.0
    # via
    #   ecdsa
    #   python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
sqlalchemy==1.4.43
    # via
    #   alembic
    #   data-inclusion-api (setup.py)
starlette==0.20.4
    # via fastapi
typing-extensions==4.4.0
    # via pydantic
urllib3==1.26.12
    # via
    #   requests
    #   sentry-sdk
uvicorn[standard]==0.19.0
    # via data-inclusion-api (setup.py)
uvloop==0.17.0
    # via uvicorn
watchfiles==0.18.1
    # via uvicorn
websockets==10.4
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
